# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "swagger 接口文档",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://www.baidu.com"
        },
        "version": "2.0"
    },
    "host": "192.168.1.163:7001",
    "paths": {
...
        "/v1/withdrawal/account_list": {
            "get": {
                "description": "获取股东提现申请列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提现"
                ],
                "summary": "获取股东提现申请列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/withdrawal/application": {
            "post": {
                "description": "创建提现申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提现"
                ],
                "summary": "创建提现申请",
                "parameters": [
                    {
                        "description": "创建提现申请 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.ApplicationWithdrawal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/withdrawal/detail": {
            "get": {
                "description": "获取提现申请详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提现"
                ],
                "summary": "获取提现申请详情",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/withdrawal/list": {
            "get": {
                "description": "获取提现申请数据列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提现"
                ],
                "summary": "获取提现申请数据列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id 可选参数",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "获取当页条数",
                        "name": "num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/withdrawal/pass": {
            "post": {
                "description": "后台通过提现申请 1、申请状态改为通过 2、需要扣除对应股东余额",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提现"
                ],
                "summary": "后台通过提现申请",
                "parameters": [
                    {
                        "description": "后台通过提现申请 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.Detail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/withdrawal/reject": {
            "post": {
                "description": "后台驳回提现申请 1、申请状态改为驳回 2、需要添加驳回原因",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提现"
                ],
                "summary": "后台驳回提现申请",
                "parameters": [
                    {
                        "description": "后台驳回提现申请 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.WithdrawalReject"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.ApplicationWithdrawal": {
            "type": "object",
            "required": [
                "amount",
                "hash",
                "shareholder_id",
                "withdrawal_pwd"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "content": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "shareholder_id": {
                    "type": "string"
                },
                "with_log": {
                    "type": "string"
                },
                "withdrawal_pwd": {
                    "type": "string"
                }
            }
        },
...
        "request.Detail": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "request.WithdrawalReject": {
            "type": "object",
            "required": [
                "content",
                "id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "response.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}